refactor: add i18n, to tsx, module css
