+++ i18n/en/code.json
@@ -363,5 +363,26 @@
   },
   "leaderboard.invalidDataAlert": {
     "message": "Invalid JSON input, please check your input"
+  },
+  "communityLeaderboard.leaderboard.title": {
+    "message": "OpenRank leaderboard"
+  },
+  "communityLeaderboard.leaderboard.login": {
+    "message": "Login"
+  },
+  "communityLeaderboard.details.title": {
+    "message": "Details"
+  },
+  "communityLeaderboard.details.from": {
+    "message": "From"
+  },
+  "communityLeaderboard.details.ratio": {
+    "message": "Ratio"
+  },
+  "communityLeaderboard.details.value": {
+    "message": "Value"
+  },
+  "communityLeaderboard.banner.title": {
+    "message": "Community OpenRank Leaderboard"
   }
 }
