+++ i18n/en/docusaurus-theme-classic/navbar.json
@@ -26,5 +26,9 @@
   "item.label.devDocs": {
     "message": "Developer Docs",
     "description": "Navbar item with label devDocs"
+  },
+  "item.label.community_openrank_leaderboard": {
+    "message": "Community OpenRank Leaderboard",
+    "description": "Navbar item with label leaderboard"
   }
 }
