--- src/components/DevLeaderboard/devLeaderboard.css
@@ -1,118 +1,0 @@
-#main {
-
-}
-
-#spin {
-  position: absolute;
-  top: 20%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.spin-box {
-  padding: 40px;
-}
-
-#graph {
-  position: absolute;
-  left: 0;
-  top: 70px;
-  width: 100%;
-  height: calc(100vh - 100px);
-}
-
-.graph-container {
-  width: 800px;
-}
-
-#control {
-  width: 320px;
-  height: 800px;
-}
-
-#list {
-  height: 300px;
-  margin: 10px;
-}
-
-#details {
-  width: 360px;
-  height: 300px;
-  margin: 10px;
-  overflow-y: scroll;
-}
-
-#leaderboard_table {
-  width: 95%;
-  margin: 10px;
-}
-
-#details_table {
-  width: 95%;
-  margin: 10px;
-}
-
-.bordered {
-  border: 2px solid grey;
-}
-
-#leaderboard_div {
-  height: 290px;
-}
-
-#details_div {
-  height: 290px;
-}
-
-.scrollit {
-  overflow: scroll;
-}
-
-table th {
-  text-align: left;
-}
-
-.banner-box {
-  z-index: 3;
-  position: absolute;
-  top: 80px;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 660px;
-  height: 120px;
-  background-color: rgba(255, 255, 255, 0.5);
-  border: 1px solid black;
-  border-radius: 1rem;
-  padding: 1rem;
-}
-
-.banner-title {
-  margin-bottom: 0.5rem;
-  text-align: center;
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: black;
-  text-shadow: 0 0 10px white;
-}
-
-.banner-content {
-  display: flex;
-  justify-content: space-between;
-}
-
-.left-box {
-  position: absolute;
-  top: 220px;
-  left: 0;
-}
-
-.leaderboard-box {
-  height: 300px;
-}
-
-.right-box {
-  position: absolute;
-  top: 220px;
-  right: 0;
-}
-
