+++ src/components/Leaderboard/styles.module.css
@@ -7,6 +7,8 @@
   margin: 0px;
   padding: 10px;
   color: #EDF9FF;
+  overflow-y: scroll;
+  height: 100%;
 }
 
 .boardTitle {
