+++ src/pages/community_openrank_leaderboard.tsx
@@ -1,0 +1,12 @@
+import Layout from '@theme/Layout';
+import DevLeaderboard from '../components/DevLeaderboard';
+
+export default (): JSX.Element => {
+  return (
+    <div className='no-footer'>
+      <Layout>
+        <DevLeaderboard />
+      </Layout>
+    </div>
+  );
+}
