--- src/pages/leaderboard.mdx
@@ -1,4 +1,0 @@
-import MetricChartComponent from '@site/src/components/OpenDiggerMetricCharts';
-import DevLeaderboard from '@site/src/components/DevLeaderboard';
-
-<DevLeaderboard />
